+++ src/functionalTests/cogrpc.test.ts
@@ -18,8 +18,10 @@ import { expectBuffersToEqual } from '../testUtils/buffers';
 import { arrayToAsyncIterable, asyncIterableToArray } from '../testUtils/iter';
 import { getPromiseRejection } from '../testUtils/jest';
 import { ExternalPdaChain, generateCCA, generateCDAChain } from '../testUtils/pki';
-import { GW_COGRPC_HOST_URL, GW_INTERNET_ADDRESS, GW_POHTTP_HOST_URL } from './services';
-import { connectToNatsStreaming, createAndRegisterPrivateGateway, sleep } from './utils';
+import { createAndRegisterPrivateGateway } from './utils/gatewayRegistration';
+import { GW_COGRPC_HOST_URL, GW_INTERNET_ADDRESS, GW_POHTTP_HOST_URL } from './utils/constants';
+import { connectToNatsStreaming } from './utils/nats';
+import { sleep } from './utils/timing';
 
 const TOMORROW = addDays(new Date(), 1);
