+++ src/functionalTests/utils/constants.ts
@@ -1,6 +1,10 @@
+import { get as getEnvVar } from 'env-var';
+
 export const GW_INTERNET_ADDRESS = 'public-gateway-pohttp.default';
 export const GW_POHTTP_HOST_URL = 'http://127.0.0.1:8081';
 export const GW_COGRPC_HOST_URL = 'https://127.0.0.1:8082';
 export const GW_POWEB_HOST_PORT = 8080;
 
 export const PONG_LOCAL_URL = 'http://127.0.0.1:8083';
+
+export const IS_GITHUB = getEnvVar('IS_GITHUB').asBool();
