+++ src/functionalTests/utils/timing.ts
@@ -1,0 +1,3 @@
+export async function sleep(seconds: number): Promise<void> {
+  return new Promise((resolve) => setTimeout(resolve, seconds * 1_000));
+}
